Manage buffer with yaml in OCaml part of ocaml-yaml
This fixes a bug when a string with YAML was passed to
yaml_parser_set_input_string, which would only store a pointer to it (not a
copy). The temporary buffer, created by the FFI from string on calling this
function was freed, and later overwritten - before being parsed by YAML parser.

This changes the signature of yaml_parser_set_input_string to accept char ptr,
and the char array is stored in parser type structure as long as is needed for
parsing.